Initialize English Directory.
This will be used in the AI engine to determine how to answer questions
or have the ability to know what is the alphabet, nouns, verbs,
adjectives and etcetera in the English language. This will use a mixture
of object oriented and procedural programming to make the program
functional and worth using. May also use other libraries to achieve
task.