Initialize Native IO Board Header.
The Native IO Board header will be used for the processing of the main
input in the terminal. This will consist of linear searches and many
more algorithms to determine what is being entered into the machine.
This will also determine what will leave the machine as output to the
user. The future development of Native will consist of adding either
open source speech recognition algorithms created by others or the
creation of self made speech recognition algorithms. This will allow for
a more AI like structure to take place.